refactor(core): reorg SAML app related codes
