+++ packages/core/src/queries/saml-application/configs.ts
