+++ packages/core/src/saml-application/SamlApplication/consts.ts
