+++ packages/core/src/saml-application/SamlApplication/utils.ts
